% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{blm}
\alias{blm}
\title{Bayesian linear model}
\usage{
blm(Y, X, S, prior)
}
\arguments{
\item{Y}{dependent variables, n by m}

\item{X}{independent variables, n by k}

\item{S}{number of draws}

\item{prior}{list of priors \eqn{(B_0,V_0,\Lambda_0,\nu_0)}}
}
\description{
Bayesian linear regression with normal-inverse-Wishart conjugate prior
\deqn{
Y|B,\Sigma,X \sim MN(XB,I,\Sigma),\
B|\Sigma \sim MN(B_0, V_0, \Sigma),\
\Sigma \sim IW(\Lambda_0, \nu_0)
}
where \eqn{MN} is the matrix normal distribution.
}
