% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{blm_cpp}
\alias{blm_cpp}
\title{Bayesian linear model}
\arguments{
\item{Y}{dependent variables}

\item{X}{independent variables}

\item{S}{sample size}

\item{prior}{list of priors}
}
\description{
Bayesian linear regression with normal-inverse-Wishart conjugate prior
prior: A ~ MN(A_0, V_0, Sigma), Sigma ~ IW(Lambda_0, nu_0)
}
