% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btse-package.R
\docType{package}
\name{btse-package}
\alias{btse}
\alias{btse-package}
\title{btse: Bayesian Time Series Econometrics}
\description{
Implementations of many Bayesian time series models for Econometrics research.
}
\author{
\strong{Maintainer}: Xiaolei Wang \email{adamwang15@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
