% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar.R
\name{bvar}
\alias{bvar}
\title{Bayesian vector autoregression}
\usage{
bvar(Y, k, S, prior)
}
\arguments{
\item{Y}{dependent variables, n by m}

\item{k}{number of lags}

\item{S}{number of draws}

\item{prior}{list of priors \eqn{(B_0,V_0,\Lambda_0,\nu_0)}}
}
\description{
Bayesian vector autoregression
}
