% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{identify}
\alias{identify}
\title{Structural shocks identification}
\usage{
identify(posterior, strategy, sign = NULL)
}
\arguments{
\item{posterior}{draws from bvar}

\item{strategy}{identification strategy, short/long/sign}

\item{sign}{sign matrix, only used when strategy is sign}
}
\description{
Needed for IRF estimation.
}
