% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{minnesota_prior}
\alias{minnesota_prior}
\title{Minnesota prior}
\usage{
minnesota_prior(
  Y,
  k,
  non_stationary = NULL,
  model = "conjugate",
  lambda_1 = 1/.Machine$double.eps,
  lambda_2 = 0.2,
  lambda_3 = 1
)
}
\arguments{
\item{Y}{data matrix}

\item{k}{lag order}

\item{model}{type of model, either "natural" or "independent"}

\item{lambda_1}{hyper parameter of variance of intercept}

\item{lambda_2}{hyper parameter variance of coefficients}

\item{lambda_3}{hyper parameter decay rate of coefficients}

\item{prior}{list of priors, might contain "non_stationary" for custom shrinkage}
}
\description{
Given data and lag order, construct Minnesota prior for Bayesian VAR.
}
