% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar.R
\name{bsvar}
\alias{bsvar}
\title{Bayesian Structural vector autoregression}
\usage{
bsvar(Y, p, S, prior, identification, sign = NULL)
}
\arguments{
\item{Y}{dependent variables, n by m}

\item{p}{number of lags}

\item{S}{number of draws}

\item{prior}{list of priors \eqn{(B_0,V_0,\Lambda_0,\nu_0)}}
}
\description{
Bayesian Structural vector autoregression
}
