% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{irf}
\alias{irf}
\title{Impulse response function}
\usage{
irf(posterior, shock_names = NULL)
}
\arguments{
\item{posterior}{posterior draws with impulse response function}

\item{shock_names}{custom shock names in order}
}
\description{
Given posterior draws from bsvar model, plot the impulse response function.
}
