% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_irf}
\alias{plot_irf}
\title{Impulse response function}
\usage{
plot_irf(posterior, variable_names)
}
\arguments{
\item{posterior}{posterior draws with impulse response function}
}
\description{
Given posterior draws from SVAR model, plot the impulse response function
}
