% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar_irf.R
\name{plot_irf}
\alias{plot_irf}
\title{Plot impulse response function}
\usage{
plot_irf(irf, variable_names, shock_names)
}
\arguments{
\item{irf}{draws of impulse response function}

\item{variable_names}{variable names}

\item{shock_names}{structural shock names}
}
\description{
Given draws of irf, plot the impulse response function.
}
